Paste text into the text editor.
With this new update you now do control - V to paste text.
It uses Xclip.
The install.sh file will try to download xclip on distrobution that are:
Arch Linux
Ubuntu
Fedora. Will update more in the future. 
So the paste feature works when you have the xclip installed on your computer.