New feature - open files with control + O
This function allows for the user to open the files they want to open without leaving the application.

Easy and simple to use. Just enter the file you want to open then if not save you can save or not and then you can open the file. 

So it will tell you if you wanna save or not for incase the user does not want to save or if they do but forgotten. 

Thanks Jonathan Steadman. :)